+++ app/src/main/java/com/bestapp/rice/ui/meetupmap/tool.kt
@@ -1,0 +1,35 @@
+package com.bestapp.rice.ui.meetupmap
+
+import android.content.Context
+import android.graphics.Bitmap
+import android.graphics.drawable.BitmapDrawable
+import coil.ImageLoader
+import coil.request.ImageRequest
+import coil.request.SuccessResult
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.withContext
+import kotlin.math.*
+
+suspend fun toBitmap(context: Context, uri: String): Bitmap? {
+    return withContext(Dispatchers.IO) {
+        val loader = ImageLoader(context)
+        val request = ImageRequest.Builder(context)
+            .data(uri)
+            .build()
+
+        val result = (loader.execute(request) as? SuccessResult)?.drawable
+        (result as? BitmapDrawable)?.bitmap
+    }
+}
+
+fun haversine(lat1: Double, lon1: Double, lat2: Double, lon2: Double): Double {
+    val R = 6371.0 // 지구 반지름 (킬로미터)
+    val dLat = Math.toRadians(lat2 - lat1)
+    val dLon = Math.toRadians(lon2 - lon1)
+    val a = sin(dLat / 2) * sin(dLat / 2) +
+            cos(Math.toRadians(lat1)) * cos(Math.toRadians(lat2)) *
+            sin(dLon / 2) * sin(dLon / 2)
+    val c = 2 * atan2(sqrt(a), sqrt(1 - a))
+    return R * c // 거리 (킬로미터)
+}
+
